change class layout
