+++ library.properties
@@ -1,5 +1,5 @@
 name=Adafruit ADS1X15
-version=1.1.2
+version=2.0.0
 author=Adafruit
 maintainer=Adafruit <info@adafruit.com>
 sentence=Arduino library for ADS1015/1115 ADCs.
